Control.External

Description

Definition of external tasks

Documentation

data InputPath

Set of items which may be treated as an input path to an external task.

Constructors

IPItem Item
An item in the content store.

IPExternalFile ExternallyAssuredFile
An external file whose contents are considered assured by the external system.

IPExternalDir ExternallyAssuredDirectory
An external directory whose contents are considered assured by the external system.

data ParamField

Component of a parameter

Constructors

ParamText !Text
Text component.

ParamPath !InputPath
Reference to a path to a content store item.

ParamEnv !Text
Reference to an environment variable.

ParamUid
Reference to the effective user ID of the executor.

ParamGid
Reference to the effective group ID of the executor.

ParamOut
Reference to the output path in the content store.

ParamCmd Param
A quoted command that we can pass to another program as an argument.

newtype Param

A parameter to an external task

The runtime values to external references, e.g. environment variables, should not significantly influence the result of the external task. In particular, the content hash will not depend on these runtime values.

Constructors

Param [ParamField]

data ConvParam f

Converter of path components.

Constructors

ConvParam

Fields

convPath :: Item -> f (Path Abs Dir)

Resolve a reference to a content store item.

convEnv :: Text -> f Text

Resolve an environment variable.

convUid :: f CUid

Resolve the effective user ID.

convGid :: f CGid

Resolve the effective group ID.

convOut :: f (Path Abs Dir)

Resolve the output path in the content store.

paramFieldToText :: Applicative f => ConvParam f -> ParamField -> f Text

paramToText :: Applicative f => ConvParam f -> Param -> f Text

Transform a parameter to text using the given converter.

stringParam :: String -> Param

textParam :: Text -> Param

pathParam :: InputPath -> Param

Reference to a path to either: - a content store item, or - an externally assured file/directory.

contentParam :: Content t -> Param

Reference to a path to a file or directory within a store item.

externalFileParam :: ExternallyAssuredFile -> Param

Reference an externally assured file

externalDirectoryParam :: ExternallyAssuredDirectory -> Param

Reference an externally assured file

envParam :: Text -> Param

Reference to an environment variable.

uidParam :: Param

Reference to the effective user ID of the executor.

gidParam :: Param

Reference to the effective group ID of the executor.

outParam :: Param

Reference to the output path in the content store.

data OutputCapture

Control how and where stdout from the process is captured. Some external steps will write their output to stdout rather than to a file.

Constructors

NoOutputCapture
Specify that the step will write its output files directly, and that stdout will not be captured in the step output.

StdOutCapture
Capture output to a file named out in the output directory.

CustomOutCapture (Path Rel File)
Capture output to a custom named file in the output directory.

outputCaptureToRelFile :: OutputCapture -> Maybe (Path Rel File)

Get the file to write output to, if this is desired.

data Env

Control the environment set for the external process. This can either inherit from the surrounding environment, or explicitly set things.

Constructors

EnvInherit
Inherit all environment variables from the surrounding shell. Note that the values of these variables will not be taken into account in the content hash, and so changes to them will not trigger a rerun of the step.

EnvExplicit [(Text, Param)]

data ExternalTask

A monomorphic description of an external task. This is basically just a command which can be run.

Constructors

ExternalTask

Fields

_etCommand :: Text

_etParams :: [Param]

Environment variables to set for the scope of the execution.

_etEnv :: Env

_etWriteToStdOut :: OutputCapture

etCommand :: Lens' ExternalTask Text

etParams :: Lens' ExternalTask [Param]

etEnv :: Lens' ExternalTask Env

etWriteToStdOut :: Lens' ExternalTask OutputCapture

data TaskDescription

Constructors

TaskDescription

Fields

_tdOutput :: ContentHash

_tdTask :: ExternalTask

tdOutput :: Lens' TaskDescription ContentHash

tdTask :: Lens' TaskDescription ExternalTask