Control.External.Coordinator.SQLite

Description

SQLLite co-ordinator for Funflow.

This co-ordinator effectively uses the shared filesystem as a tool for task distribution and sequencing. This means that it can control a distributed funflow task without needing any additional processes running.

Documentation

data SQLite

SQLite coordinator tag.

Constructors

SQLite