Control.External.Executor

Description

Executor for external tasks.

An executor will poll for tasks on the co-ordinator, mark them as in progress and then execute them.

You probably want to start with executeLoop.

Documentation

data ExecutionResult

Constructors

Cached
The result already exists in the store and there is no need to execute. This is also returned if the job is already running elsewhere.

AlreadyRunning
The computation is already running elsewhere. This is probably indicative of a bug, because the coordinator should only allow one instance of a task to be running at any time.

Success TimeSpec
Execution completed successfully after a certain amount of time.

Failure TimeSpec Int
Execution failed with the following exit code. TODO where should logs go?

ExecutorFailure IOException
The executor itself failed to execute the external task. E.g. because the executable was not found.

execute :: ContentStore -> TaskDescription -> KatipContextT IO ExecutionResult

Execute an individual task.

executeLoop :: forall c. Coordinator c => c -> Config c -> ContentStore -> IO ()

Execute tasks forever

executeLoopWithScribe :: forall c. Coordinator c => c -> Config c -> ContentStore -> Scribe -> IO ()

Same as executeLoop, but allows specifying a custom Scribe for logging

withFollowFile :: Path Abs File -> Handle -> IO a -> IO a

withFollowFile in out follows the file in and prints contents to out as they appear. The file must exist. Doesn't handle file truncation.